voice_service/requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asteroid-filterbanks==0.4.0
attrs==25.1.0
av==14.2.0
babel==2.17.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.115.11
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
HyperPyYAML==1.2.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
kokoro==0.8.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lightning==2.5.0.post0
lightning-utilities==0.13.1
loguru==0.7.3
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
misaki==0.8.4
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
num2words==0.5.14
numpy==1.26.4
omegaconf==2.3.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
phonemizer-fork==3.3.2
pillow==11.1.0
preshed==3.0.9
primePy==1.3
propcache==0.3.0
protobuf==6.30.0
psycopg2==2.9.10
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segments==2.3.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.2
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.18.3
librosa==0.11.0